更换了语义分析思路
不再创建分析树。更新了算术表达式，不过目前()里面不能有赋值。

+++ SymbolTable.h
@@ -86,7 +86,7 @@ private:
 	map<string, FuncNode*> funcTable;
 
 	TypeNode bTypes[4] = {
-		{"int", 0},{ "int", 0 },{ "int", 0 },{ "int", 0 }
+		{"int", 0},{ "float", 0 },{ "int", 0 },{ "int", 0 }
 	};
 
 };
